requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
MLB-StatsAPI==1.8.1
requests==2.32.3
urllib3==2.3.0
